# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LeadStage = typeof LeadStage[keyof typeof LeadStage];


export const LeadStage = {
  novo: 'novo',
  qualificacao: 'qualificacao',
  fatura: 'fatura',
  proposta: 'proposta',
  documentos: 'documentos',
  assinatura: 'assinatura',
  ativacao: 'ativacao',
  ganho: 'ganho',
  perdido: 'perdido',
} as const;

export type CustomerType = typeof CustomerType[keyof typeof CustomerType];


export const CustomerType = {
  CPF: 'CPF',
  CNPJ: 'CNPJ',
} as const;

export type BrazilState = typeof BrazilState[keyof typeof BrazilState];


export const BrazilState = {
  PE: 'PE',
  CE: 'CE',
} as const;

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  nota: 'nota',
  ligacao: 'ligacao',
  whatsapp: 'whatsapp',
  email: 'email',
  mudanca_etapa: 'mudanca_etapa',
} as const;

export interface Lead {
  id: number;
  name: string;
  phone: string;
  customerType: CustomerType;
  cpfCnpj: string;
  state: BrazilState;
  city: string;
  distributor: string;
  averageBill: number;
  estimatedMonthlySavings: number;
  estimatedAnnualSavings: number;
  source: string;
  stage: LeadStage;
  consentAt: string;
  consentText: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  nextFollowUpAt?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface LeadInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  phone: string;
  customerType: CustomerType;
  /** @minLength 5 */
  cpfCnpj: string;
  state: BrazilState;
  /** @minLength 2 */
  city: string;
  /** @minLength 2 */
  distributor: string;
  /** @minimum 0 */
  averageBill: number;
  /** @minimum 0 */
  estimatedMonthlySavings: number;
  /** @minimum 0 */
  estimatedAnnualSavings: number;
  /** @minLength 1 */
  source: string;
  consentAccepted: true;
  /** @minLength 10 */
  consentText: string;
}

export interface LeadUpdate {
  /** @minLength 2 */
  name?: string;
  /** @minLength 8 */
  phone?: string;
  stage?: LeadStage;
  /** @nullable */
  nextFollowUpAt?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Activity {
  id: number;
  leadId: number;
  type: ActivityType;
  content: string;
  /** @nullable */
  authorId?: string | null;
  createdAt: string;
}

export interface ActivityInput {
  type: ActivityType;
  /** @minLength 1 */
  content: string;
}

export interface FollowUpTask {
  id: number;
  leadId: number;
  title: string;
  dueAt: string;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface FollowUpTaskInput {
  /** @minLength 1 */
  title: string;
  dueAt: string;
  assignedTo?: string;
}

export interface FollowUpTaskUpdate {
  /** @minLength 1 */
  title?: string;
  dueAt?: string;
  completed?: boolean;
  assignedTo?: string;
}

export type LeadDetail = Lead & {
  activities: Activity[];
  tasks: FollowUpTask[];
};

export interface DashboardKpis {
  totalLeads: number;
  newLeads: number;
  activeLeads: number;
  wonLeads: number;
  lostLeads: number;
  monthlySavings: number;
  dueFollowUps: number;
}

export interface Error {
  error: string;
}

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

export type ListLeadsParams = {
search?: string;
stage?: LeadStage;
state?: BrazilState;
customerType?: CustomerType;
};