# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityType } from './activityType';

export interface Activity {
  id: number;
  leadId: number;
  type: ActivityType;
  content: string;
  /** @nullable */
  authorId?: string | null;
  createdAt: Date;
}