# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityType } from './activityType';

export interface ActivityInput {
  type: ActivityType;
  /** @minLength 1 */
  content: string;
}