# lib/api-zod/src/generated/types/activityType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  nota: 'nota',
  ligacao: 'ligacao',
  whatsapp: 'whatsapp',
  email: 'email',
  mudanca_etapa: 'mudanca_etapa',
} as const;