# lib/api-zod/src/generated/types/customerType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */

export type CustomerType = typeof CustomerType[keyof typeof CustomerType];


export const CustomerType = {
  CPF: 'CPF',
  CNPJ: 'CNPJ',
} as const;