# lib/api-zod/src/generated/types/followUpTask.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface FollowUpTask {
  id: number;
  leadId: number;
  title: string;
  dueAt: Date;
  completed: boolean;
  /** @nullable */
  completedAt?: Date | null;
  /** @nullable */
  assignedTo?: string | null;
  createdAt: Date;
  updatedAt: Date;
}