# lib/api-zod/src/generated/types/followUpTaskInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface FollowUpTaskInput {
  /** @minLength 1 */
  title: string;
  dueAt: Date;
  assignedTo?: string;
}