# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { BrazilState } from './brazilState';
import type { CustomerType } from './customerType';
import type { LeadStage } from './leadStage';

export interface Lead {
  id: number;
  name: string;
  phone: string;
  customerType: CustomerType;
  cpfCnpj: string;
  state: BrazilState;
  city: string;
  distributor: string;
  averageBill: number;
  estimatedMonthlySavings: number;
  estimatedAnnualSavings: number;
  source: string;
  stage: LeadStage;
  consentAt: Date;
  consentText: string;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  nextFollowUpAt?: Date | null;
  /** @nullable */
  notes?: string | null;
}