# lib/api-zod/src/generated/types/leadDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { FollowUpTask } from './followUpTask';
import type { Lead } from './lead';

export type LeadDetail = Lead & {
  activities: Activity[];
  tasks: FollowUpTask[];
};