# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { BrazilState } from './brazilState';
import type { CustomerType } from './customerType';

export interface LeadInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  phone: string;
  customerType: CustomerType;
  /** @minLength 5 */
  cpfCnpj: string;
  state: BrazilState;
  /** @minLength 2 */
  city: string;
  /** @minLength 2 */
  distributor: string;
  /** @minimum 0 */
  averageBill: number;
  /** @minimum 0 */
  estimatedMonthlySavings: number;
  /** @minimum 0 */
  estimatedAnnualSavings: number;
  /** @minLength 1 */
  source: string;
  consentAccepted: true;
  /** @minLength 10 */
  consentText: string;
}