# lib/api-zod/src/generated/types/leadStage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */

export type LeadStage = typeof LeadStage[keyof typeof LeadStage];


export const LeadStage = {
  novo: 'novo',
  qualificacao: 'qualificacao',
  fatura: 'fatura',
  proposta: 'proposta',
  documentos: 'documentos',
  assinatura: 'assinatura',
  ativacao: 'ativacao',
  ganho: 'ganho',
  perdido: 'perdido',
} as const;