# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PYX Energia CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStage } from './leadStage';

export interface LeadUpdate {
  /** @minLength 2 */
  name?: string;
  /** @minLength 8 */
  phone?: string;
  stage?: LeadStage;
  /** @nullable */
  nextFollowUpAt?: Date | null;
  /** @nullable */
  notes?: string | null;
}